debian/patches/debianization.patch: new
Keep local symbols local.

This is necessary to avoid clashes between functions  with the same names
in pam_ssh and other applications.
One known example is buffer_append(), that is also in dovecot.

One of the consequences in such cases is pam_ssh to fail/crash.

based on the libpam_ssh package by Jerome Benoit <calculus@rezozer.net>